+++ src/main/groovy/com/toastcoders/vmware/yavijava/contracts/YavijavaHTMLClientAbs.groovy
@@ -1,5 +1,6 @@
-package com.toastcoders.vmware.yavijava
+package com.toastcoders.vmware.yavijava.contracts
 
+import com.toastcoders.vmware.yavijava.YavijavaDataObjectHTMLClient
 import com.toastcoders.vmware.yavijava.contracts.HTMLClient
 import org.apache.log4j.Logger
 import org.jsoup.Jsoup
